feat: remove magic link auth example in favor of OTP auth in expo example
