+++ examples/react-native-expo-example/app.json
@@ -26,6 +26,16 @@
 		"web": {
 			"favicon": "./assets/favicon.png"
 		},
-		"plugins": ["expo-router"]
+		"plugins": [
+			"expo-router",  
+			[
+				"expo-build-properties",
+				{
+					"ios": {
+						"deploymentTarget": "17.0"
+					}
+				}
+			]
+		]
 	}
 }
