+++ examples/react-native-expo-example/package.json
@@ -10,11 +10,12 @@
 		"start:redirect-server": "tsx ./redirect-server/index.ts"
 	},
 	"dependencies": {
-		"@account-kit/react": "4.15.0",
-		"@account-kit/react-native": "4.15.0",
-		"@account-kit/react-native-signer": "4.15.0",
-		"@account-kit/signer": "4.15.0",
+		"@account-kit/react": "4.16.0",
+		"@account-kit/react-native": "4.16.0",
+		"@account-kit/react-native-signer": "4.16.0",
+		"@account-kit/signer": "4.16.0",
 		"expo": "~52.0.18",
+		"expo-build-properties": "~0.13.2",
 		"expo-constants": "~17.0.3",
 		"expo-linking": "~7.0.3",
 		"expo-router": "4.0.11",
