+++ site/pages/reference/account-kit/react/functions/createConfig.mdx
@@ -1,27 +1,59 @@
 ---
 # This file is autogenerated
+
 title: createConfig
 description: Overview of the createConfig method
 ---
 
 # createConfig
 
-Creates an AlchemyAccountsConfig by using the provided parameters to configure the core settings, including the required transport. It includes a signer creation function internally.
+Wraps the `createConfig` that is exported from `@aa-sdk/core` to allow passing
+an additional argument, the configuration object for the Auth Components UI
+(the modal and AuthCard).
 
 ## Import
 
 ```ts
-import { createConfig } from "@account-kit/react-native";
+import { createConfig } from "@account-kit/react";
+```
+
+## Usage
+
+```ts
+import { sepolia, alchemy } from "@account-kit/infra"
+import { AlchemyAccountsUIConfig, createConfig } from "@account-kit/react"
+import { QueryClient } from "@tanstack/react-query";
+
+const uiConfig: AlchemyAccountsUIConfig = {
+illustrationStyle: "linear",
+auth: {
+sections: [[{ type: "email" }], [{ type: "passkey" }]],
+addPasskeyOnSignup: true,
+},
+}
+
+const config = createConfig({
+transport: alchemy({ apiKey: "your_api_key" })
+chain: sepolia,
+ssr: true,
+}, uiConfig)
+
+export const queryClient = new QueryClient();
 ```
 
 ## Parameters
 
-### params
+### props
 
-`BaseCreateConfigProps`
-The base properties required for creating the config and establishing client store settings
+`CreateConfigProps`
+for creating an alchemy account config
+
+### ui
+
+`AlchemyAccountsUIConfig`
+(optional) configuration to use for the Auth Components UI
 
 ## Returns
 
-`AlchemyAccountsConfig`
-An object of type AlchemyAccountsConfig containing the configured properties
+`AlchemyAccountsConfigWithUI`
+an alchemy account config object containing the core and client store, as well as the UI config
